+++ lib/jwt_decoder.dart
@@ -51,37 +51,48 @@ class JwtDecoder {
   /// Throws [FormatException] if parameter is not a valid JWT token.
   static bool isExpired(String token) {
     final expirationDate = getExpirationDate(token);
+    if (expirationDate == null) {
+      return false;
+    }
     // If the current date is after the expiration date, the token is already expired
     return DateTime.now().isAfter(expirationDate);
+  }
+
+  static DateTime? _getDate({required String token, required String claim}) {
+    final decodedToken = decode(token);
+    final expiration = decodedToken[claim] as int?;
+    if (expiration == null) {
+      return null;
+    }
+    return DateTime.fromMillisecondsSinceEpoch(expiration * 1000);
   }
 
   /// Returns token expiration date
   ///
   /// Throws [FormatException] if parameter is not a valid JWT token.
-  static DateTime getExpirationDate(String token) {
-    final decodedToken = decode(token);
-
-    final expirationDate = DateTime.fromMillisecondsSinceEpoch(0)
-        .add(Duration(seconds: decodedToken['exp'].toInt()));
-    return expirationDate;
+  static DateTime? getExpirationDate(String token) {
+    return _getDate(token: token, claim: 'exp');
   }
 
   /// Returns token issuing date (iat)
   ///
   /// Throws [FormatException] if parameter is not a valid JWT token.
-  static Duration getTokenTime(String token) {
-    final decodedToken = decode(token);
-
-    final issuedAtDate = DateTime.fromMillisecondsSinceEpoch(0)
-        .add(Duration(seconds: decodedToken["iat"]));
+  static Duration? getTokenTime(String token) {
+    final issuedAtDate = _getDate(token: token, claim: 'iat');
+    if (issuedAtDate == null) {
+      return null;
+    }
     return DateTime.now().difference(issuedAtDate);
   }
 
   /// Returns remaining time until expiry date.
   ///
   /// Throws [FormatException] if parameter is not a valid JWT token.
-  static Duration getRemainingTime(String token) {
+  static Duration? getRemainingTime(String token) {
     final expirationDate = getExpirationDate(token);
+    if (expirationDate == null) {
+      return null;
+    }
 
     return expirationDate.difference(DateTime.now());
   }